build_agent/requirements.txt
langchain
langchain-community
langchain-openai
streamlit
python-dotenv
mcp
httpx
